[TASK] Allow floats for TCA type number range values
If the field is using `format=decimal` the range should be allowed to be
defined as floats as well.

Resolves: #101812
Releases: main, 13.4, 12.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88036
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/backend/Classes/Form/Element/NumberElement.php
@@ -190,8 +190,8 @@ class NumberElement extends AbstractFormElement
             $rangeAttributes = [
                 'type' => 'range',
                 'class' => 'slider',
-                'min' => (string)(int)($config['range']['lower'] ?? 0),
-                'max' => (string)(int)($config['range']['upper'] ?? 10000),
+                'min' => (string)(float)($config['range']['lower'] ?? 0),
+                'max' => (string)(float)($config['range']['upper'] ?? 10000),
                 'step' => (string)($config['slider']['step'] ?? 1),
                 'style' => 'width: ' . (int)($config['slider']['width'] ?? 400) . 'px',
                 'title' => (string)$itemValue,
@@ -216,10 +216,10 @@ class NumberElement extends AbstractFormElement
         $resultArray = $this->mergeChildReturnIntoExistingResult($resultArray, $fieldWizardResult, false);
 
         if (isset($config['range']['lower'])) {
-            $attributes['min'] = (string)(int)$config['range']['lower'];
+            $attributes['min'] = (string)(float)$config['range']['lower'];
         }
         if (isset($config['range']['upper'])) {
-            $attributes['max'] = (string)(int)$config['range']['upper'];
+            $attributes['max'] = (string)(float)$config['range']['upper'];
         }
 
         if ($format === 'decimal') {
